worker/requirements.txt
boto3>=1.35.0
python-dotenv>=1.0.0
playwright>=1.49.0
pypdf>=5.1.0
Pillow>=10.0.0
img2pdf>=0.5.0
pytest>=8.0.0
